삭제 확인 창 추가, 오류 수정

+++ src/main/webapp/WEB-INF/views/readPost.jsp
@@ -94,9 +94,9 @@
 			
 			        <!-- 게시글 관련 버튼 -->
 			        <c:if test="${sessionScope.login.id == post.writer}">
-			            <a href="../post/list?page=${page}" role="button" class="btn btn-info bi bi-arrow-return-left" id=listbtn> 목록</a>
+			            <a href="javascript:window.history.back();" role="button" class="btn btn-info bi bi-arrow-return-left" id=listbtn> 목록</a>
 			            <a href="../post/update?id=${post.id}" role="button" class="btn btn-primary bi bi-pencil-square" id=updatebtn> 수정</a>
-			            <button type="button" onclick="location.href='../post/delete?id=${post.id}'" id="btn-delete" class="btn btn-danger bi bi-trash" id=deletebtn> 삭제</button>
+			            <button type="button" onclick="if(!confirm('게시글을 정말 삭제하시겠습니까?')){return false;} location.href='../post/delete?id=${post.id}'" id="btn-delete" class="btn btn-danger bi bi-trash" id=deletebtn> 삭제</button>
 			        </c:if>
 			        <c:if test="${sessionScope.login.id != post.writer}">
 			            <a href="javascript:window.history.back();" role="button" class="btn btn-info bi bi-arrow-return-left" id=listbtn> 목록</a>
